backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Auth & DB
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Database
sqlalchemy==2.0.36

# Firebase Admin (for OAuth sync)
firebase-admin==6.6.0

# Stripe
stripe==11.3.0

# Azure Blob Storage
azure-storage-blob==12.19.0

# AI
anthropic==0.39.0

# Env
python-dotenv==1.0.1
pydantic-settings==2.6.1
